refactor(extract): pull out the 'parse' step from etract.js ...
... rename all modules that extract dependencies from an AST from extract-<modulesystem> to etract-<modulesystem>-deps as that better conveys their purpose

... put the extract-yadda-deps in a separate folder

... separate parsing and extracting in the (new) typescript dependency extraction
