+++ src/extract/ast-extractors/extract-commonJS-deps.js
