+++ src/extract/extract.js
@@ -1,61 +1,15 @@
 "use strict";
 
-const fs                          = require('fs');
 const path                        = require('path');
-const acorn                       = require('acorn');
-const acorn_loose                 = require('acorn/dist/acorn_loose');
 const _                           = require('lodash');
 const resolve                     = require('./resolve');
-const transpile                   = require('./transpile');
 const ignore                      = require('./ignore');
-const extractES6Dependencies      = require('./extract-ES6');
-const extractCommonJSDependencies = require('./extract-commonJS');
-const extractAMDDependencies      = require('./extract-AMD');
-const extractTypeScript           = require('./extract-typescript');
-
-/**
- * Returns the extension of the given file name path.
- *
- * Just using path.extname would be fine for most cases,
- * except for coffeescript, where a markdown extension can
- * mean literate coffeescript.
- *
- * @param {string} pFileName path to the file to be parsed
- * @return {string}          extension
- */
-function getExtension(pFileName) {
-    let lRetval = path.extname(pFileName);
-
-    if (lRetval === ".md") {
-        return pFileName.endsWith(".coffee.md") ? ".coffee.md" : lRetval;
-    }
-    return lRetval;
-}
-
-/**
- * Returns the abstract syntax tree of the module identified by the passed
- * file name. It obtains this by (1) transpiling the contents of the file
- * into javascript (if necessary) (2) parsing it.
- *
- * If parsing fails we fall back to acorn's 'loose' parser
- *
- * @param {string} pFileName path to the file to be parsed
- * @returns {object}         the abstract syntax tree
- */
-function getAST(pFileName) {
-    const lFile = transpile(
-        getExtension(pFileName),
-        fs.readFileSync(pFileName, 'utf8')
-    );
-
-    try {
-        return acorn.parse(lFile, {sourceType: 'module'});
-    } catch (e) {
-        return acorn_loose.parse_dammit(lFile, {sourceType: 'module'});
-    }
-}
-const getASTCached = _.memoize(getAST);
-
+const extractES6Dependencies      = require('./ast-extractors/extract-ES6-deps');
+const extractCommonJSDependencies = require('./ast-extractors/extract-commonJS-deps');
+const extractAMDDependencies      = require('./ast-extractors/extract-AMD-deps');
+const extractTypeScript           = require('./ast-extractors/extract-typescript-deps');
+const getJSASTCached              = require('./parse/toJavascriptAST').getASTCached;
+const toTypescriptAST             = require('./parse/toTypescriptAST');
 
 /**
  * Returns an array of dependencies present in the given file. Of
@@ -91,7 +45,7 @@ module.exports = (pFileName, pOptions) => {
             pOptions
         );
 
-        const lAST = getASTCached(path.join(lOptions.baseDir, pFileName));
+        const lAST = getJSASTCached(path.join(lOptions.baseDir, pFileName));
         let lDependencies = [];
 
         if (lOptions.moduleSystems.indexOf("cjs") > -1){
@@ -99,9 +53,13 @@ module.exports = (pFileName, pOptions) => {
         }
 
         if (lOptions.moduleSystems.indexOf("es6") > -1) {
-            if (getExtension(pFileName).startsWith(".ts")) {
+            if (path.extname(pFileName).startsWith(".ts") && toTypescriptAST.isAvailable()) {
                 lDependencies = lDependencies.concat(
-                    extractTypeScript(fs.readFileSync(path.join(lOptions.baseDir, pFileName), 'utf8'))
+                    extractTypeScript(
+                        toTypescriptAST.getASTCached(
+                            path.join(lOptions.baseDir, pFileName)
+                        )
+                    )
                 );
             } else {
                 extractES6Dependencies(lAST, lDependencies);
