+++ test/extract/ast-extractors/extract-typescript-commonjs.spec.js
@@ -1,0 +1,18 @@
+const expect            = require('chai').expect;
+const extractTypescript = require('./extract-typescript.utl');
+
+describe("extract-typescript - regular commonjs require", () => {
+
+    it("extracts require of a module that uses an export-equals'", () => {
+        expect(
+            extractTypescript("import thing = require('./thing-that-uses-export-equals');")
+        ).to.deep.equal(
+            [
+                {
+                    moduleName: './thing-that-uses-export-equals',
+                    moduleSystem: 'cjs'
+                }
+            ]
+        );
+    });
+});
