+++ test/extract/ast-extractors/extract-typescript.utl.js
@@ -1,0 +1,5 @@
+
+const extractTypescript = require('../../../src/extract/ast-extractors/extract-typescript-deps');
+const getASTFromSource  = require('../../../src/extract/parse/toTypescriptAST').getASTFromSource;
+
+module.exports = pTypesScriptSource => extractTypescript(getASTFromSource(pTypesScriptSource));
